requirements.txt
flask==3.0.2
flask-cors==4.0.0
flask-jwt-extended==4.6.0
flask-pymongo==2.3.0
pymongo==4.6.2
python-dotenv==1.0.1
bcrypt==4.1.2
cloudinary==1.39.0
twilio==8.13.0
Werkzeug==3.0.1
dnspython==2.6.1
